+++ src/main/mainApp.vue
@@ -10,9 +10,14 @@
 
 <script>
 import { mapActions, mapGetters, mapState } from 'vuex';
+import { v4 as uuidv4 } from 'uuid';
 import '@formatjs/intl-numberformat/polyfill';
 import '@formatjs/intl-numberformat/locale-data/en';
 
+import {
+  getInjectedName,
+  getInjectedIcon
+} from '@/core/helpers/detectProvider.js';
 import { PWA_EVENTS } from '@/core/helpers/common';
 import {
   Toast,
@@ -65,6 +70,22 @@ export default {
     });
   },
   mounted() {
+    // manually add web3 wallet detected
+    if (window.ethereum) {
+      const name = getInjectedName(window.ethereum);
+      const info = {
+        rdns: 'com.detected.injectedwallet',
+        uuid: uuidv4(),
+        name: name,
+        icon: getInjectedIcon(name)
+      };
+      const provider = window.ethereum;
+      this.storeEIP6963Wallet({ info, provider });
+    }
+    // epi6963 listener
+    window.addEventListener('eip6963:announceProvider', e => {
+      this.storeEIP6963Wallet(e.detail);
+    });
     EventBus.$on('swapTxBroadcasted', hash => {
       this.trackSwap('swapTxBroadcasted', hash, this.network.type.chainID);
     });
@@ -127,6 +148,7 @@ export default {
     EventBus.$off('swapTxNotBroadcastedFailed');
     document.removeEventListener('visibilitychange');
     window.removeEventListener('mouseout');
+    window.removeEventListener('eip6963:announceProvider');
   },
   methods: {
     ...mapActions('global', ['setOnlineStatus']),
@@ -135,6 +157,7 @@ export default {
     ...mapActions('article', ['updateArticles']),
     ...mapActions('article', ['updateArticles']),
     ...mapActions('popups', ['showSurveyPopup']),
+    ...mapActions('external', ['storeEIP6963Wallet']),
     openBuy() {
       this.buySellOpen = true;
     },