+++ src/modules/network/ModuleNetwork.vue
@@ -62,6 +62,7 @@ export default {
   computed: {
     ...mapState('wallet', ['blockNumber', 'identifier', 'isHardware']),
     ...mapState('global', ['validNetwork']),
+    ...mapState('external', ['selectedEIP6963Provider']),
     ...mapGetters('global', ['network']),
     fullName() {
       return this.network.type.name_long;
@@ -74,15 +75,15 @@ export default {
     },
     show() {
       let switchNetworkWeb3Supported = false;
-      if (window.ethereum) {
+      if (this.selectedEIP6963Provider) {
         const isMetaMask =
-          window.ethereum.isMetaMask &&
-          !window.ethereum.hasOwnProperty('isTrust') &&
-          !window.ethereum.hasOwnProperty('isMEWwallet');
+          this.selectedEIP6963Provider.isMetaMask &&
+          !this.selectedEIP6963Provider.hasOwnProperty('isTrust') &&
+          !this.selectedEIP6963Provider.hasOwnProperty('isMEWwallet');
         const isMEWwallet =
-          window.ethereum.isMetaMask &&
-          window.ethereum.isMEWwallet &&
-          window.ethereum.isTrust;
+          this.selectedEIP6963Provider.isMetaMask &&
+          this.selectedEIP6963Provider.isMEWwallet &&
+          this.selectedEIP6963Provider.isTrust;
         switchNetworkWeb3Supported = isMetaMask || isMEWwallet;
       }