+++ src/utils/web3-provider/methods/eth_sendTransaction.js
@@ -28,7 +28,7 @@ export default async ({ payload, store, requestManager }, res, next) => {
     };
   }
   let currency = store.getters['external/contractToToken'](tx.to);
-  if (!currency)
+  if (!(currency.name && currency.symbol))
     currency = store.getters['external/contractToToken'](MAIN_TOKEN_ADDRESS);
   tx.gasPrice = tx.gasPrice
     ? tx.gasPrice