+++ src/views/layouts-default/TheAccessWalletLayout.vue
@@ -48,6 +48,7 @@
             NOT RECOMMENDED
           </div>
           <mew-button
+            v-if="btn.useBtn"
             has-full-width
             :class="[
               btn.title === 'Software'
@@ -89,6 +90,56 @@
               </div>
             </div>
           </mew-button>
+          <div v-else class="non-button-container mb-5 py-6">
+            <div class="width--full d-flex align-center text-left">
+              <img
+                v-if="btn.icon && !isMobile"
+                class="ml-5 mr-6"
+                :src="btn.icon"
+                :alt="btn.alt"
+                style="height: 70px"
+              />
+              <div class="px-3">
+                <div class="d-flex align-center">
+                  <img
+                    v-if="btn.icon && isMobile"
+                    class="mr-4"
+                    :src="btn.icon"
+                    :alt="btn.alt"
+                    style="height: 40px"
+                  />
+
+                  <div class="mew-heading-2 break-word letter-spacing--initial">
+                    {{ btn.title }}
+                  </div>
+                </div>
+                <div
+                  class="mew-heading-4 reset-subtitle break-word letter-spacing--initial text-transform--none mt-2 mb-2"
+                >
+                  {{ btn.subtitle }}
+                </div>
+                <div
+                  v-if="eip6963Providers.length > 1"
+                  class="d-flex align-center justify-start mini-button-container"
+                >
+                  <div
+                    v-for="item in eip6963Providers"
+                    :key="item.info.uuid"
+                    class="mr-2 px-1 py-2 d-flex align-center cursor--pointer mini-buttons"
+                    @click="openWeb3WithProvider(item)"
+                  >
+                    <img
+                      v-if="item.info.icon"
+                      :src="item.info.icon"
+                      :alt="`${item.info.name}-picture`"
+                      width="25px"
+                    />
+                    <span class="pl-1">{{ item.info.name }}</span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
         </div>
       </div>
 
@@ -165,7 +216,7 @@ export default {
     };
   },
   computed: {
-    ...mapState('external', ['path']),
+    ...mapState('external', ['path', 'eip6963Providers']),
     ...mapState('wallet', ['isOfflineApp']),
     /**
      * Opens up software module overlay. Returns true if overlay prop from route is ACCESS_VALID_OVERLAYS.SOFTWARE
@@ -205,9 +256,11 @@ export default {
             subtitle: 'Connect with Enkrypt browser extension',
             official: true,
             recommended: true,
+            useBtn: true,
             icon: require('@/assets/images/icons/icon-enkrypt-block.svg'),
             alt: 'Enkrypt',
             fn: () => {
+              this.trackAccessWalletAmplitude('click_access_enkrypt');
               this.checkEnkrypt();
             }
           },
@@ -218,9 +271,11 @@ export default {
             subtitle: 'Connect MEW Wallet app to MEW web',
             official: true,
             recommended: true,
+            useBtn: true,
             icon: require('@/assets/images/icons/icon-mew-wallet.png'),
             alt: 'MEW wallet',
             fn: () => {
+              this.trackAccessWalletAmplitude('click_access_mew_wallet');
               this.openMEWwallet();
             }
           },
@@ -231,9 +286,11 @@ export default {
             subtitle: 'Use your Web3 wallet with MEW',
             official: false,
             recommended: true,
+            useBtn: this.eip6963Providers.length <= 1,
             icon: require('@/assets/images/icons/icon-extensions.png'),
             alt: 'Hardware Wallets',
             fn: () => {
+              this.trackAccessWalletAmplitude('click_access_browser_extension');
               this.openWeb3Wallet();
             }
           },
@@ -244,9 +301,13 @@ export default {
             subtitle: 'WalletConnect, WalletLink',
             official: false,
             recommended: true,
+            useBtn: true,
             icon: require('@/assets/images/icons/icon-mobile-apps.png'),
             alt: 'Hardware Wallets',
             fn: () => {
+              this.trackAccessWalletAmplitude(
+                'click_access_open_mobile_wallet'
+              );
               this.openOverlay(ACCESS_VALID_OVERLAYS.MOBILE);
             }
           },
@@ -257,9 +318,13 @@ export default {
             subtitle: 'Ledger, Trezor, KeepKey, Cool Wallet, Bitbox02',
             official: false,
             recommended: true,
+            useBtn: true,
             icon: require('@/assets/images/icons/icon-hardware-wallet.png'),
             alt: 'Hardware Wallets',
             fn: () => {
+              this.trackAccessWalletAmplitude(
+                'click_access_open_hardware_wallet'
+              );
               this.openOverlay(ACCESS_VALID_OVERLAYS.HARDWARE);
             }
           },
@@ -270,8 +335,12 @@ export default {
             title: 'Software',
             subtitle: 'Keystore File, Mnemonic Phrase, and Private Key',
             official: false,
+            useBtn: true,
             recommended: false,
             fn: () => {
+              this.trackAccessWalletAmplitude(
+                'click_access_open_software_wallet'
+              );
               this.openOverlay(ACCESS_VALID_OVERLAYS.SOFTWARE);
             }
           }
@@ -281,8 +350,10 @@ export default {
         {
           color: 'white',
           title: 'Software',
+          useBtn: true,
           subtitle: 'Keystore files, Mnemonic phrase, Private key',
           fn: () => {
+            this.trackAccessWalletAmplitude('click_access_open_mobile_wallet');
             this.openOverlay(ACCESS_VALID_OVERLAYS.SOFTWARE);
           }
         }
@@ -292,8 +363,15 @@ export default {
       return this.$vuetify.breakpoint.smAndDown;
     }
   },
+  mounted() {
+    window.dispatchEvent(new Event('eip6963:requestProvider'));
+  },
   methods: {
     ...mapActions('wallet', ['setWallet']),
+    ...mapActions('external', [
+      'setSelectedEIP6963Info',
+      'setSelectedEIP6963Provider'
+    ]),
     /**
      * Pushes route to empty Access wallet with no props
      * Consequently closing any open overlay
@@ -301,6 +379,7 @@ export default {
      */
     close() {
       try {
+        this.trackAccessWalletAmplitude(`close_access_${this.overlay}_wallet`);
         this.$router.push({
           name: ROUTES_HOME.ACCESS_WALLET.NAME
         });
@@ -308,11 +387,16 @@ export default {
         Toast(e, {}, ERROR);
       }
     },
+    openWeb3WithProvider(item) {
+      this.trackAccessWalletAmplitude('click_access_browser_extension');
+      this.openWeb3Wallet(item);
+    },
     openMEWwallet() {
       try {
         WalletConnectWallet(WALLET_TYPES.MEW_WALLET)
           .then(_newWallet => {
             this.setWallet([_newWallet]).then(() => {
+              this.trackAccessWalletAmplitude('access_wallet_success');
               this.trackAccessWallet(WALLET_TYPES.MEW_WALLET);
               this.$router.push({ name: ROUTES_WALLET.DASHBOARD.NAME });
             });
@@ -345,6 +429,15 @@ export default {
      * Checks if Enkrypt is available
      */
     checkEnkrypt() {
+      if (this.eip6963Providers.length > 0) {
+        const item = this.eip6963Providers.find(item => {
+          if (item.info.name.toLowerCase() === 'enkrypt') return item;
+        });
+        if (item) {
+          this.openWeb3Wallet(item);
+          return;
+        }
+      }
       if (
         window.ethereum &&
         window.ethereum.isMetaMask &&
@@ -358,15 +451,20 @@ export default {
     /**
      * Checks and open web3 wallet
      */
-    async openWeb3Wallet(provider) {
-      if (provider || window.ethereum) {
-        const providedProvider = provider || window.ethereum;
+    async openWeb3Wallet(item) {
+      if (item || window.ethereum) {
+        if (item) {
+          this.setSelectedEIP6963Info(item.info);
+          this.setSelectedEIP6963Provider(item.provider);
+        }
+        const providedProvider = item ? item.provider : window.ethereum;
         const web3 = new Web3(providedProvider);
         try {
           await providedProvider.enable();
           const acc = await web3.eth.requestAccounts();
           const wallet = new Web3Wallet(acc[0]);
           this.setWallet([wallet, providedProvider]);
+          this.trackAccessWalletAmplitude('access_wallet_success');
           this.trackAccessWallet(WALLET_TYPES.WEB3_WALLET);
           if (this.path !== '') {
             this.$router.push({ path: this.path });
@@ -418,5 +516,29 @@ export default {
   right: 0px;
   padding: 4px 8px;
   border-radius: 0px 10px 0 7px;
+}
+.non-button-container {
+  background-color: white;
+  border-radius: 10px;
+  padding: 24px 20px;
+  letter-spacing: 0.5px;
+  text-transform: none;
+  width: 100%;
+  height: initial;
+  min-height: 157px;
+  align-items: center;
+  display: flex;
+}
+
+.mini-buttons {
+  border: 1px solid var(--v-greenPrimary-base);
+  border-radius: 4px;
+  width: 110px;
+  height: 45px;
+}
+
+.mini-button-container {
+  width: 100%;
+  overflow-x: auto;
 }
 </style>